requirements.txt
numpy~=1.25.0
opencv-python~=4.8.0.74
torch~=2.0.1
pillow~=10.0.0
diffusers~=0.27.2